Rename chat_* files to youtube_*
This so so other mods that use the chat_* files do not conflict with the HUD. Luckily I do not plan to replace any strings in this HUD, so we can use youtube_* without consequence, hopefully. Realistically it shouldn't cause any problems, since the YouTube functionality has been long deprecated.